requirements.txt
Package                   Version
------------------------- -----------
altair                    5.5.0
attrs                     25.1.0
blinker                   1.9.0
cachetools                5.5.1
certifi                   2024.12.14
charset-normalizer        3.4.1
click                     8.1.8
colorama                  0.4.6
dnspython                 2.7.0
et_xmlfile                2.0.0
faiss-cpu                 1.9.0.post1
filelock                  3.17.0
fsspec                    2024.12.0
gitdb                     4.0.12
GitPython                 3.1.44
huggingface-hub           0.28.1
idna                      3.10
Jinja2                    3.1.5
joblib                    1.4.2
jsonschema                4.23.0
jsonschema-specifications 2024.10.1
markdown-it-py            3.0.0
MarkupSafe                3.0.2
mdurl                     0.1.2
mpmath                    1.3.0
narwhals                  1.24.1
networkx                  3.4.2
numpy                     2.2.2
openpyxl                  3.1.5
packaging                 24.2
pandas                    2.2.3
pillow                    11.1.0
pip                       22.3.1
protobuf                  5.29.3
pyarrow                   19.0.0
pydeck                    0.9.1
Pygments                  2.19.1
pymongo                   10.10.10.10
python-dateutil           2.9.0.post0
pytz                      2024.2
PyYAML                    6.0.2
referencing               0.36.2
regex                     2024.11.6
requests                  2.32.3
rich                      13.9.4
rpds-py                   0.22.3
safetensors               0.5.2
scikit-learn              1.6.1
scipy                     1.15.1
sentence-transformers     3.4.1
setuptools                65.5.0
six                       1.17.0
smmap                     5.0.2
streamlit                 1.41.1
sympy                     1.13.1
tenacity                  9.0.0
threadpoolctl             3.5.0
tokenizers                0.21.0
toml                      0.10.2
torch                     2.6.0
tornado                   6.4.2
tqdm                      4.67.1
transformers              4.48.1
typing_extensions         4.12.2
tzdata                    2025.1
urllib3                   2.3.0
watchdog                  6.0.0
